+++ tests/AlphaFS.UnitTest/UnitTest Utility/TemporaryDirectory.cs
@@ -1,4 +1,4 @@
-/*  Copyright (C) 2008-2018 Peter Palotas, Jeffrey Jangli, Alexandr Normuradov
+﻿/*  Copyright (C) 2008-2018 Peter Palotas, Jeffrey Jangli, Alexandr Normuradov
  *  
  *  Permission is hereby granted, free of charge, to any person obtaining a copy 
  *  of this software and associated documentation files (the "Software"), to deal 
@@ -261,13 +261,13 @@ namespace AlphaFS.UnitTest
          {
             dirSecurity = Alphaleonis.Win32.Filesystem.Directory.GetAccessControl(folderFullPath);
             dirSecurity.AddAccessRule(rule);
-            Alphaleonis.Win32.Filesystem.Directory.SetAccessControl(folderFullPath, dirSecurity);
+            Alphaleonis.Win32.Filesystem.Directory.SetAccessControl(folderFullPath, dirSecurity, AccessControlSections.Access);
          }
          else
          {
             dirSecurity = Alphaleonis.Win32.Filesystem.Directory.GetAccessControl(folderFullPath);
             dirSecurity.RemoveAccessRule(rule);
-            Alphaleonis.Win32.Filesystem.Directory.SetAccessControl(folderFullPath, dirSecurity);
+            Alphaleonis.Win32.Filesystem.Directory.SetAccessControl(folderFullPath, dirSecurity, AccessControlSections.Access);
          }
       }